Add heuristic that simplifies the input
The heuristic is based on the observation that for each(generator, chip)
pair on the 1st floor the time to solve increases by 12; and for each
pair on the 2nd floor the time increases by 8 (so in general 4
additional steps per pair per floor .. however for the 3rd floor this
is not quite so simple - in general it is true that additional pairs
cost 4 there, but for small values there are some anomalies .. so avoid
optimizing that). So we can remove those pairs from the problem and
just add the time to the result.